[CofenseTriage] Move all classes into one file
The plugin architecture requires it.

on-behalf-of: @Cofense <oss@cofense.com>
